Move rule support modules under rules/support
