+++ src/rules/key_duplicates.rs
@@ -1,7 +1,7 @@
 use saphyr_parser::{Event, Parser, Span, SpannedEventReceiver};
 
 use crate::config::YamlLintConfig;
-use crate::rules::mapping_key_walker::Walker;
+use crate::rules::support::mapping_key_walker::Walker;
 
 pub const ID: &str = "key-duplicates";
