+++ src/rules/mod.rs
@@ -10,17 +10,15 @@ pub mod document_start;
 pub mod empty_lines;
 pub mod empty_values;
 pub mod float_values;
-pub(crate) mod flow_collection;
 pub mod hyphens;
 pub mod indentation;
 pub mod key_duplicates;
 pub mod key_ordering;
 pub mod line_length;
-pub(crate) mod mapping_key_walker;
 pub mod new_line_at_end_of_file;
 pub mod new_lines;
 pub mod octal_values;
 pub mod quoted_strings;
-pub(crate) mod span_utils;
+pub(crate) mod support;
 pub mod trailing_spaces;
 pub mod truthy;
