+++ src/rules/quoted_strings.rs
@@ -3,7 +3,7 @@ use saphyr::Yaml;
 use saphyr_parser::{Event, Parser, ScalarStyle, Span, SpannedEventReceiver, Tag};
 
 use crate::config::YamlLintConfig;
-use crate::rules::mapping_key_walker::Walker;
+use crate::rules::support::mapping_key_walker::Walker;
 
 pub const ID: &str = "quoted-strings";
