+++ src/rules/support/mapping_key_walker.rs
