+++ src/rules/support/span_utils.rs
